# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chainsmith API",
    "description": "Headless Rust API for the Chainsmith Flesh and Blood deck builder.",
    "license": {
      "name": "PolyForm-Noncommercial-1.0.0",
      "identifier": "PolyForm-Noncommercial-1.0.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/decks": {
      "get": {
        "operationId": "listDecks",
        "responses": {
          "200": {
            "description": "Decks owned by the authenticated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Authorization header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "createDeck",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Deck created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Authorization header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/decks/{id}": {
      "get": {
        "operationId": "getDeck",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Deck id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full deck",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Authorization header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Deck not found or not visible to caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "deleteDeck",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Deck id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted (soft)"
          },
          "401": {
            "description": "Missing or invalid Authorization header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Deck not found or not owned by caller",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateDeckRequest": {
        "type": "object",
        "required": [
          "name",
          "deck"
        ],
        "properties": {
          "deck": {
            "$ref": "#/components/schemas/Deck",
            "description": "The pool/hero/format/loadouts to save."
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string",
            "description": "Display name shown to the user (and to other users for public decks)."
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "visibility": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Visibility"
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "Deck": {
        "type": "object",
        "required": [
          "format",
          "hero",
          "pool",
          "loadouts"
        ],
        "properties": {
          "format": {
            "$ref": "#/components/schemas/FormatId"
          },
          "hero": {
            "$ref": "#/components/schemas/PrintingId"
          },
          "loadouts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Loadout"
            }
          },
          "pool": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PoolEntry"
            }
          }
        },
        "additionalProperties": false
      },
      "DeckListResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeckSummary"
            }
          }
        }
      },
      "DeckResponse": {
        "type": "object",
        "required": [
          "id",
          "ownerId",
          "name",
          "visibility",
          "tags",
          "deck",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "deck": {
            "$ref": "#/components/schemas/Deck"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility"
          }
        }
      },
      "DeckSummary": {
        "type": "object",
        "required": [
          "id",
          "name",
          "format",
          "hero",
          "visibility",
          "tags",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "format": {
            "$ref": "#/components/schemas/FormatId"
          },
          "hero": {
            "$ref": "#/components/schemas/PrintingId"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility"
          }
        }
      },
...
      "EquipmentLoadout": {
        "type": "object",
        "properties": {
          "arms": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PrintingId"
              }
            ]
          },
          "chest": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PrintingId"
              }
            ]
          },
          "head": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PrintingId"
              }
            ]
          },
          "legs": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PrintingId"
              }
            ]
          },
          "mainHand": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PrintingId",
                "description": "Main-hand slot. Holds a weapon (1H or 2H). A 2H weapon here occupies\nboth hand slots; see `off_hand` for the Companion exception."
              }
            ]
          },
          "offHand": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PrintingId",
                "description": "Off-hand slot. Holds either a 1H weapon, an off-hand equipment piece\n(shield, Trade Routes, etc.), or — when `main_hand` carries a 2H\nweapon — only a Companion-subtyped off-hand equipment piece."
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "description": "Stable, machine-readable error codes. Clients are allowed to switch on\nthese. Adding a variant is additive; renaming or removing one is a\nbreaking change in production phase (see `.claude/rules/api-contract.md`).",
        "enum": [
          "database_error",
          "unsupported_format",
          "not_found",
          "unauthorized",
          "forbidden",
          "auth_not_configured"
        ]
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {},
          "message": {
            "type": "string"
          }
        }
      },
...
      "FormatId": {
        "type": "string",
        "enum": [
          "classic_constructed",
          "blitz",
          "commoner",
          "silver_age",
          "living_legend",
          "ultimate_pit_fight"
        ]
      },
...
      "Loadout": {
        "type": "object",
        "required": [
          "name",
          "deckCards",
          "equipment"
        ],
        "properties": {
          "deckCards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoadoutEntry"
            }
          },
          "equipment": {
            "$ref": "#/components/schemas/EquipmentLoadout"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "LoadoutEntry": {
        "type": "object",
        "required": [
          "printing",
          "quantity"
        ],
        "properties": {
          "printing": {
            "$ref": "#/components/schemas/PrintingId"
          },
          "quantity": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "PoolEntry": {
        "type": "object",
        "required": [
          "printing",
          "quantity"
        ],
        "properties": {
          "printing": {
            "$ref": "#/components/schemas/PrintingId"
          },
          "quantity": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
...
      "PrintingId": {
        "type": "string"
      },
...
      "Visibility": {
        "type": "string",
        "enum": [
          "private",
          "unlisted",
          "public"
        ]
      },
...
    }
  }
}